wip: #71; entry files get watched/reprocessed properly
