+++ lib/core/codegen.js
@@ -12,6 +12,7 @@ function getProps(cayoElement) {
   );
 }
 
+// Generate the Cayo Runtime which is unique per page to render the cayos on the page
 export function generateCayoRuntime(components, fileName, _cayo) {
   const { config, __VERSION__ } = _cayo;
   let code = '';
@@ -80,4 +81,9 @@ export function generateComponentInstance(cayoId, componentName) {
   });
 `
   );
+}
+
+// Generate the code to import the the generated Cayo Runtime file
+export function generateCayoRuntimeImport() {
+  return `import { default as renderCayos } from './cayo-runtime.js';\n`
 }
