+++ lib/core/dependencies.js
@@ -138,3 +138,31 @@ export function findDependentCayos(pageDependency, _cayo) {
 
   return cayos;
 }
+
+export async function getEntryDependencies(entry, page, _cayo) {
+  const { config } = _cayo;
+  let imports = precinct(entry.code, { type: 'es6' });
+  // Filter out public path ('/') and node_modules ('<package>')
+  let localDeps = imports.filter(d => d.startsWith('../') || d.startsWith('./'));
+  let absoluteDeps = localDeps.map(d => path.resolve(path.dirname(entry.path), d));
+  let srcRelativeLocalDeps = localDeps.map((d, i) => {
+    let pageCayoPath = path.resolve(config.cayoPath, page.url === '/' ? './' : page.url);
+    return path.relative(pageCayoPath, absoluteDeps[i]);
+  });
+  // let nodeModules = imports
+  //   .filter(d => !d.startsWith('../') && !d.startsWith('./') && !d.startsWith('/'))
+  //   .map(d => path.relative(config.cayoPath, `${config.projectRoot}/node_modules/${d}`));
+  
+  entry.dependencies = [];
+  for (let i = 0; i < localDeps.length; i++) {
+    entry.dependencies.push([localDeps[i], srcRelativeLocalDeps[i]]);
+  }
+
+
+  const depender = { 
+    type: 'entry', 
+    path: entry.path, 
+    dependencies: absoluteDeps, 
+  }
+  await handleDependencies(depender, _cayo);
+}
