+++ lib/core/files.js
@@ -2,6 +2,7 @@ import fs from 'fs-extra';
 import path from 'path';
 import chalk from 'chalk';
 import { logger } from './logger.js';
+import { processEntrySource } from './entry.js';
 
 // Write file content for a page
 // TODO: wrap all fs stuff in try blocks (need the awaits?)
@@ -56,12 +57,10 @@ export async function writeEntryFile(page, _cayo) {
   const { config } = _cayo;
   const { entry } = page.result;
   
-  if (entry.path && entry.code) {
-    let code = `${entry.code}`;
-    for (const [relativeDep, srcRelativeDep] of entry.dependencies) {
-      code = code.replace(relativeDep, srcRelativeDep);
-    }
-
+  if (entry.path) {
+    // Get the processed entry code
+    let code = processEntrySource(entry, page, _cayo);
+    // Write the processed entry file
     let entryOutputPath = page.url === '/' ? `./index.js` : `${page.url}/index.js`;
     await fs.outputFile(path.resolve(config.cayoPath, entryOutputPath), code)
       .then(() => logger.log.info(
