+++ lib/core/render/prerender.js
@@ -7,9 +7,6 @@ import { logger } from '../logger.js';
 import { generateCayoRuntime } from '../codegen.js';
 import { generateCayoComponentId } from '../utils.js';
 import { compileCayos } from '../compile/cayos.js';
-import { getDeps } from '../bundle.js';
-import { handleDependencies } from '../dependencies.js';
-import precinct from 'precinct';
 
 export async function prerender(page, _cayo) {
   const renderer = new Renderer(page.layout);
@@ -144,15 +141,8 @@ export async function processPage(content, page, _cayo, logger) {
           entryScriptPlaceholder.remove();
         }
       } else {
-        try {
-          const entrySource = await fs.readFile(entryFilePath, { encoding: 'utf8' });
-          if (entrySource) entry.code = entrySource;
-          entry.path = entryFilePath;
-          _cayo.stats.dependencies.pages[page.sourcePath].add(entry.path);
-          handleEntryDependencies(page, entry, _cayo);
-        } catch (err) {
-          throw err;
-        }
+        entry.path = entryFilePath;
+        _cayo.stats.dependencies.pages[page.sourcePath].add(entry.path);
       }
 
     } else {
@@ -185,10 +175,9 @@ export async function processPage(content, page, _cayo, logger) {
   const cayoRuntimePath = page.url === '/' ? `/index.js` : `${page.url}/index.js`;
   js = generateCayoRuntime(cayoInstances, cayoRuntimePath, _cayo);
 
-  // Prepend cayo runtime import to entry source code
   if (cayoIds.length > 0) {
-    entry.code = `import { default as renderCayos } from './cayo-runtime.js';\n` 
-      + entry.code;
+    // Indicate that we need to make this entry ready to render cayos later
+    entry.renderCayos = true;
   }
   
   // Handle runtime warnings  
@@ -313,32 +302,4 @@ function generateRuntimeIssuesScript(runtime, issues, type) {
   }
 
   return script;
-}
-
-async function handleEntryDependencies(page, entry, _cayo) {
-  const { config } = _cayo;
-  let imports = precinct(entry.code, { type: 'es6', includeCore: false });
-  // Filter out public path ('/') and node_modules ('<package>')
-  let localDeps = imports.filter(d => d.startsWith('../') || d.startsWith('./'));
-  let absoluteDeps = localDeps.map(d => path.resolve(path.dirname(entry.path), d));
-  let srcRelativeLocalDeps = localDeps.map((d, i) => {
-    let pageCayoPath = path.resolve(config.cayoPath, page.url === '/' ? './' : page.url);
-    return path.relative(pageCayoPath, absoluteDeps[i]);
-  });
-  // let nodeModules = imports
-  //   .filter(d => !d.startsWith('../') && !d.startsWith('./') && !d.startsWith('/'))
-  //   .map(d => path.relative(config.cayoPath, `${config.projectRoot}/node_modules/${d}`));
-  
-  entry.dependencies = [];
-  for (let i = 0; i < localDeps.length; i++) {
-    entry.dependencies.push([localDeps[i], srcRelativeLocalDeps[i]]);
-  }
-
-
-  const depender = { 
-    type: 'entry', 
-    path: entry.path, 
-    dependencies: absoluteDeps, 
-  }
-  await handleDependencies(depender, _cayo);
 }
